测试数据库+充值记录

+++ SPLSPOJO/src/main/java/com/zl/pojo/TopUpRecord.java
@@ -13,7 +13,7 @@ public class TopUpRecord implements Serializable {
     
 	private static final long serialVersionUID = -8760818574711753324L;
 	//主键
-	private Integer id;
+	private Long id;
 	//充值记录编号
     private String topUpNo;
     //充值时间
@@ -23,9 +23,9 @@ public class TopUpRecord implements Serializable {
     //充值金额
     private BigDecimal amount;
     //充值操作用户主键
-    private Integer userId;
+    private Long userId;
 
-    public TopUpRecord(Integer id, String topUpNo, Date createTime, String method, BigDecimal amount, Integer userId) {
+    public TopUpRecord(Long id, String topUpNo, Date createTime, String method, BigDecimal amount, Long userId) {
         this.id = id;
         this.topUpNo = topUpNo;
         this.createTime = createTime;
@@ -38,11 +38,11 @@ public class TopUpRecord implements Serializable {
         super();
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -78,11 +78,11 @@ public class TopUpRecord implements Serializable {
         this.amount = amount;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 }
